+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Only parse pyproject.toml for config
 
 ### Removed
+- Remove default `audit` values
 - Remove default `exclude` values
 - Remove `pyaud.package`
 - Remove logging